Clean up unused variables
The variables "project" and "environment" were only used for tagging.
This limits the module to use a specific set of tags, which might not fit all use cases.
The user can still provide an arbitrary set of tags that fits their needs, via the tags variable.
I've still left the "domain" tag, as it does represent something relevant in the module.

Fixes #28